actualizacion proyecto final, se agrego las opciones de equipos y agregar equipos a la sidebar
