+++ proyectofinal-henry/src/helpers/sidebarOptions.ts
@@ -18,12 +18,7 @@ export const sidebarOptionsAdmin = [
                     {optionName: "Servicios",
                      router: true, 
                      subOptions: []},  
-                     {optionName: "Equipos",
-                        router: false,   
-                        subOptions: [{subOptionsName: "Equipos", router: true},
-                                     {subOptionsName: "Agregar Equipo", router: true}]
-                       }, 
-                     
+                   
                     ]
         },
         {
@@ -31,8 +26,25 @@ export const sidebarOptionsAdmin = [
             icon: "M8 7V6a1 1 0 0 1 1-1h11a1 1 0 0 1 1 1v7a1 1 0 0 1-1 1h-1M3 18v-7a1 1 0 0 1 1-1h11a1 1 0 0 1 1 1v7a1 1 0 0 1-1 1H4a1 1 0 0 1-1-1Zm8-3.5a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0Z",
             router: true,
             options:[],
-        }
+        },
+        {
+            rolName: "Equipos",
+            icon: "m14.304 4.844 2.852 2.852M7 7H4a1 1 0 0 0-1 1v10a1 1 0 0 0 1 1h11a1 1 0 0 0 1-1v-4.5m2.409-9.91a2.017 2.017 0 0 1 0 2.853l-6.844 6.844L8 14l.713-3.565 6.844-6.844a2.015 2.015 0 0 1 2.852 0Z",
+            router: false,
+            options:[
+                   
+                     {optionName: "Equipos",
+                        router: true,   
+                        subOptions: []
+                       }, 
+                       { optionName: "Agregar Equipo",
+                        router: true,   
+                        subOptions:[]}
+                     
+                    ]
+        },
  ]
+
